# Conflicts:
#	.appveyor.yml
#	.gitignore
#	.travis.yml
#	README.md
#	TODO.md
#	ardublockly/ardublockly.css
#	ardublockly/ardublockly.js
#	ardublockly/ardublockly_blockly.js
#	ardublockly/ardublockly_design.js
#	ardublockly/ardublockly_toolbox.xml
#	ardublockly/ardublocklyserver_ajax.js
#	ardublockly/index.html
#	ardublocklyserver/py23/__init__.py
#	ardublocklyserver/tests/actions_test.py
#	blockly/README.md
#	blockly/blockly_compressed.js
#	blockly/blockly_uncompressed.js
#	blockly/blocks/arduino/io.js
#	blockly/blocks/arduino/serial.js
#	blockly/blocks/arduino/servo.js
#	blockly/blocks/arduino/spi.js
#	blockly/blocks/arduino/stepper.js
#	blockly/blocks/arduino/variables.js
#	blockly/blocks/lists.js
#	blockly/blocks/logic.js
#	blockly/blocks/loops.js
#	blockly/blocks/math.js
#	blockly/blocks/text.js
#	blockly/blocks/variables.js
#	blockly/blocks_compressed.js
#	blockly/build.py
#	blockly/core/block.js
#	blockly/core/block_svg.js
#	blockly/core/blocks.js
#	blockly/core/bubble.js
#	blockly/core/connection.js
#	blockly/core/css.js
#	blockly/core/field.js
#	blockly/core/field_angle.js
#	blockly/core/field_variable.js
#	blockly/core/inject.js
#	blockly/core/input.js
#	blockly/core/msg.js
#	blockly/core/mutator.js
#	blockly/core/static_typing.js
#	blockly/core/toolbox.js
#	blockly/core/utils.js
#	blockly/core/widgetdiv.js
#	blockly/core/workspace.js
#	blockly/core/workspace_svg.js
#	blockly/demos/blockfactory/factory.js
#	blockly/demos/plane/generated/ar.js
#	blockly/demos/plane/generated/be-tarask.js
#	blockly/demos/plane/generated/br.js
#	blockly/demos/plane/generated/ca.js
#	blockly/demos/plane/generated/da.js
#	blockly/demos/plane/generated/de.js
#	blockly/demos/plane/generated/el.js
#	blockly/demos/plane/generated/en.js
#	blockly/demos/plane/generated/es.js
#	blockly/demos/plane/generated/fa.js
#	blockly/demos/plane/generated/fr.js
#	blockly/demos/plane/generated/he.js
#	blockly/demos/plane/generated/hrx.js
#	blockly/demos/plane/generated/hu.js
#	blockly/demos/plane/generated/ia.js
#	blockly/demos/plane/generated/is.js
#	blockly/demos/plane/generated/it.js
#	blockly/demos/plane/generated/ja.js
#	blockly/demos/plane/generated/ko.js
#	blockly/demos/plane/generated/ms.js
#	blockly/demos/plane/generated/nb.js
#	blockly/demos/plane/generated/nl.js
#	blockly/demos/plane/generated/pl.js
#	blockly/demos/plane/generated/pms.js
#	blockly/demos/plane/generated/pt-br.js
#	blockly/demos/plane/generated/ro.js
#	blockly/demos/plane/generated/ru.js
#	blockly/demos/plane/generated/sc.js
#	blockly/demos/plane/generated/sv.js
#	blockly/demos/plane/generated/th.js
#	blockly/demos/plane/generated/tr.js
#	blockly/demos/plane/generated/uk.js
#	blockly/demos/plane/generated/vi.js
#	blockly/demos/plane/generated/zh-hans.js
#	blockly/demos/plane/generated/zh-hant.js
#	blockly/generators/arduino.js
#	blockly/generators/arduino/boards.js
#	blockly/generators/arduino/io.js
#	blockly/generators/arduino/servo.js
#	blockly/generators/arduino/spi.js
#	blockly/generators/arduino/stepper.js
#	blockly/generators/arduino/text.js
#	blockly/generators/arduino/variables.js
#	blockly/generators/php.js
#	blockly/generators/php/lists.js
#	blockly/generators/php/loops.js
#	blockly/generators/php/math.js
#	blockly/generators/php/text.js
#	blockly/media/sprites.png
#	blockly/msg/js/ar.js
#	blockly/msg/js/az.js
#	blockly/msg/js/bcc.js
#	blockly/msg/js/be-tarask.js
#	blockly/msg/js/bg.js
#	blockly/msg/js/bn.js
#	blockly/msg/js/br.js
#	blockly/msg/js/ca.js
#	blockly/msg/js/cs.js
#	blockly/msg/js/da.js
#	blockly/msg/js/de.js
#	blockly/msg/js/el.js
#	blockly/msg/js/en.js
#	blockly/msg/js/es.js
#	blockly/msg/js/fa.js
#	blockly/msg/js/fi.js
#	blockly/msg/js/fr.js
#	blockly/msg/js/he.js
#	blockly/msg/js/hi.js
#	blockly/msg/js/hrx.js
#	blockly/msg/js/hu.js
#	blockly/msg/js/ia.js
#	blockly/msg/js/id.js
#	blockly/msg/js/is.js
#	blockly/msg/js/it.js
#	blockly/msg/js/ja.js
#	blockly/msg/js/ko.js
#	blockly/msg/js/lb.js
#	blockly/msg/js/lrc.js
#	blockly/msg/js/lt.js
#	blockly/msg/js/mk.js
#	blockly/msg/js/ms.js
#	blockly/msg/js/nb.js
#	blockly/msg/js/nl.js
#	blockly/msg/js/oc.js
#	blockly/msg/js/pl.js
#	blockly/msg/js/pms.js
#	blockly/msg/js/pt-br.js
#	blockly/msg/js/pt.js
#	blockly/msg/js/ro.js
#	blockly/msg/js/ru.js
#	blockly/msg/js/sc.js
#	blockly/msg/js/sk.js
#	blockly/msg/js/sq.js
#	blockly/msg/js/sr.js
#	blockly/msg/js/sv.js
#	blockly/msg/js/ta.js
#	blockly/msg/js/th.js
#	blockly/msg/js/tl.js
#	blockly/msg/js/tlh.js
#	blockly/msg/js/tr.js
#	blockly/msg/js/uk.js
#	blockly/msg/js/vi.js
#	blockly/msg/js/zh-hans.js
#	blockly/msg/js/zh-hant.js
#	blockly/msg/json/cs.json
#	blockly/msg/json/da.json
#	blockly/msg/json/de.json
#	blockly/msg/json/en.json
#	blockly/msg/json/fr.json
#	blockly/msg/json/he.json
#	blockly/msg/json/hu.json
#	blockly/msg/json/id.json
#	blockly/msg/json/ko.json
#	blockly/msg/json/lt.json
#	blockly/msg/json/nl.json
#	blockly/msg/json/pl.json
#	blockly/msg/json/pt-br.json
#	blockly/msg/json/qqq.json
#	blockly/msg/json/sc.json
#	blockly/msg/json/sk.json
#	blockly/msg/json/synonyms.json
#	blockly/msg/json/ta.json
#	blockly/msg/json/tr.json
#	blockly/msg/json/uk.json
#	blockly/msg/json/vi.json
#	blockly/msg/json/zh-hant.json
#	blockly/msg/messages.js
#	blockly/php_compressed.js
#	blockly/tests/generators/index.html
#	blockly/tests/jsunit/blockly_test.js
#	blockly/tests/playground.html
#	package/README.md
#	package/build_py2exe.py
#	package/build_pyinstaller.py
#	package/electron/.gitignore
#	package/electron/README.md
#	package/electron/app/appmenu.js
#	package/electron/app/main.js
#	package/electron/app/package.json
#	package/electron/app/servermgr.js
#	package/electron/app/vendor/electron_boilerplate/context_menu.js
#	package/electron/app/vendor/electron_boilerplate/external_links.js
#	package/electron/app/vendor/electron_boilerplate/window_state.js
#	package/electron/package.json
#	package/electron/resources/linux/app.desktop
#	package/electron/resources/osx/Info.plist
#	package/electron/resources/windows/installer.nsi
#	package/electron/tasks/build.js
#	package/electron/tasks/release_linux.js
#	package/electron/tasks/release_osx.js
#	package/electron/tasks/release_windows.js
#	package/electron/tasks/utils.js
#	package/pack_ardublockly.py
#	package/pyinstaller.spec
#	package/requirements.txt

+++ blockly/core/blockly.js
@@ -27,7 +27,6 @@
 // Top level object for Blockly.
 goog.provide('Blockly');
 
-// Blockly core dependencies.
 goog.require('Blockly.BlockSvg');
 goog.require('Blockly.FieldAngle');
 goog.require('Blockly.FieldCheckbox');
@@ -42,16 +41,17 @@ goog.require('Blockly.FieldVariable');
 goog.require('Blockly.Generator');
 goog.require('Blockly.Msg');
 goog.require('Blockly.Procedures');
-goog.require('Blockly.Realtime');
 goog.require('Blockly.Toolbox');
 goog.require('Blockly.WidgetDiv');
 goog.require('Blockly.WorkspaceSvg');
 goog.require('Blockly.inject');
 goog.require('Blockly.utils');
-
-// Closure dependencies.
 goog.require('goog.color');
+goog.require('goog.userAgent');
 
+
+// Turn off debugging when compiled.
+var CLOSURE_DEFINES = {'goog.DEBUG': false};
 
 /**
  * Required name space for SVG elements.
@@ -79,8 +79,8 @@ Blockly.HSV_VALUE = 0.65;
  * Sprited icons and images.
  */
 Blockly.SPRITE = {
-  width: 64,
-  height: 92,
+  width: 96,
+  height: 124,
   url: 'sprites.png'
 };
 
@@ -89,7 +89,7 @@ Blockly.SPRITE = {
  * @param {number} hue Hue on a colour wheel (0-360).
  * @return {string} RGB code, e.g. '#5ba65b'.
  */
-Blockly.makeColour = function(hue) {
+Blockly.hueToRgb = function(hue) {
   return goog.color.hsvToHex(hue, Blockly.HSV_SATURATION,
       Blockly.HSV_VALUE * 255);
 };
@@ -207,7 +207,7 @@ Blockly.clipboardXml_ = null;
 
 /**
  * Source of the local clipboard.
- * @type {Blockly.Workspace}
+ * @type {Blockly.WorkspaceSvg}
  * @private
  */
 Blockly.clipboardSource_ = null;
@@ -248,7 +248,7 @@ Blockly.svgResize = function(workspace) {
   while (mainWorkspace.options.parentWorkspace) {
     mainWorkspace = mainWorkspace.options.parentWorkspace;
   }
-  var svg = mainWorkspace.options.svg;
+  var svg = mainWorkspace.getParentSvg();
   var div = svg.parentNode;
   if (!div) {
     // Workspace deteted, or something.
@@ -294,6 +294,9 @@ Blockly.onMouseUp_ = function(e) {
  * @private
  */
 Blockly.onMouseMove_ = function(e) {
+  if (e.touches && e.touches.length >= 2) {
+    return;  // Multi-touch gestures won't have e.clientX.
+  }
   var workspace = Blockly.getMainWorkspace();
   if (workspace.isScrolling) {
     Blockly.removeAllRanges();
@@ -311,10 +314,9 @@ Blockly.onMouseMove_ = function(e) {
 
     // Move the scrollbars and the page will scroll automatically.
     workspace.scrollbar.set(-x - metrics.contentLeft,
-                                        -y - metrics.contentTop);
+                            -y - metrics.contentTop);
     // Cancel the long-press if the drag has moved too far.
-    var dr = Math.sqrt(Math.pow(dx, 2) + Math.pow(dy, 2));
-    if (dr > Blockly.DRAG_RADIUS) {
+    if (Math.sqrt(dx * dx + dy * dy) > Blockly.DRAG_RADIUS) {
       Blockly.longStop_();
     }
     e.stopPropagation();
@@ -331,6 +333,7 @@ Blockly.onKeyDown_ = function(e) {
     // When focused on an HTML text input widget, don't trap any keys.
     return;
   }
+  var deleteBlock = false;
   if (e.keyCode == 27) {
     // Pressing esc closes the context menu.
     Blockly.hideChaff();
@@ -338,8 +341,7 @@ Blockly.onKeyDown_ = function(e) {
     // Delete or backspace.
     try {
       if (Blockly.selected && Blockly.selected.isDeletable()) {
-        Blockly.hideChaff();
-        Blockly.selected.dispose(true, true);
+        deleteBlock = true;
       }
     } finally {
       // Stop the browser from going back to the previous page.
@@ -350,14 +352,14 @@ Blockly.onKeyDown_ = function(e) {
   } else if (e.altKey || e.ctrlKey || e.metaKey) {
     if (Blockly.selected &&
         Blockly.selected.isDeletable() && Blockly.selected.isMovable()) {
-      Blockly.hideChaff();
       if (e.keyCode == 67) {
         // 'c' for copy.
+        Blockly.hideChaff();
         Blockly.copy_(Blockly.selected);
       } else if (e.keyCode == 88) {
         // 'x' for cut.
         Blockly.copy_(Blockly.selected);
-        Blockly.selected.dispose(true, true);
+        deleteBlock = true;
       }
     }
     if (e.keyCode == 86) {
@@ -365,6 +367,16 @@ Blockly.onKeyDown_ = function(e) {
       if (Blockly.clipboardXml_) {
         Blockly.clipboardSource_.paste(Blockly.clipboardXml_);
       }
+    }
+  }
+  if (deleteBlock) {
+    // Common code for delete and cut.
+    Blockly.hideChaff();
+    var heal = Blockly.dragMode_ != 2;
+    Blockly.selected.dispose(heal, true);
+    if (Blockly.highlightedConnection_) {
+      Blockly.highlightedConnection_.unhighlight();
+      Blockly.highlightedConnection_ = null;
     }
   }
 };
@@ -391,19 +403,15 @@ Blockly.longPid_ = 0;
  * which after about a second opens the context menu.  The tasks is killed
  * if the touch event terminates early.
  * @param {!Event} e Touch start event.
- * @param {Blockly.Block} block The block under the touchstart event, or null
- *   if the event was on the workspace.
+ * @param {!Blockly.Block|!Blockly.WorkspaceSvg} uiObject The block or workspace
+ *   under the touchstart event.
  * @private
  */
-Blockly.longStart_ = function(e, block) {
+Blockly.longStart_ = function(e, uiObject) {
   Blockly.longStop_();
   Blockly.longPid_ = setTimeout(function() {
       e.button = 2;  // Simulate a right button click.
-      if (block) {
-        block.onMouseDown_(e);
-      } else {
-        Blockly.onMouseDown_(e);
-      }
+      uiObject.onMouseDown_(e);
     }, Blockly.LONGPRESS);
 };
 
@@ -426,13 +434,34 @@ Blockly.longStop_ = function() {
  */
 Blockly.copy_ = function(block) {
   var xmlBlock = Blockly.Xml.blockToDom_(block);
-  Blockly.Xml.deleteNext(xmlBlock);
+  if (Blockly.dragMode_ != 2) {
+    Blockly.Xml.deleteNext(xmlBlock);
+  }
   // Encode start position in XML.
   var xy = block.getRelativeToSurfaceXY();
   xmlBlock.setAttribute('x', block.RTL ? -xy.x : xy.x);
   xmlBlock.setAttribute('y', xy.y);
   Blockly.clipboardXml_ = xmlBlock;
   Blockly.clipboardSource_ = block.workspace;
+};
+
+/**
+ * Duplicate this block and its children.
+ * @param {!Blockly.Block} block Block to be copied.
+ * @private
+ */
+Blockly.duplicate_ = function(block) {
+  // Save the clipboard.
+  var clipboardXml = Blockly.clipboardXml_;
+  var clipboardSource = Blockly.clipboardSource_;
+
+  // Create a duplicate via a copy/paste operation.
+  Blockly.copy_(block);
+  block.workspace.paste(Blockly.clipboardXml_);
+
+  // Restore the clipboard.
+  Blockly.clipboardXml_ = clipboardXml;
+  Blockly.clipboardSource_ = clipboardSource;
 };
 
 /**
@@ -482,35 +511,37 @@ Blockly.hideChaff = function(opt_allowToolbox) {
  * @this Blockly.WorkspaceSvg
  */
 Blockly.getMainWorkspaceMetrics_ = function() {
-  var svgSize = Blockly.svgSize(this.options.svg);
+  var svgSize = Blockly.svgSize(this.getParentSvg());
   if (this.toolbox_) {
     svgSize.width -= this.toolbox_.width;
   }
-  var viewWidth = svgSize.width - Blockly.Scrollbar.scrollbarThickness;
-  var viewHeight = svgSize.height - Blockly.Scrollbar.scrollbarThickness;
+  // Set the margin to match the flyout's margin so that the workspace does
+  // not jump as blocks are added.
+  var MARGIN = Blockly.Flyout.prototype.CORNER_RADIUS - 1;
+  var viewWidth = svgSize.width - MARGIN;
+  var viewHeight = svgSize.height - MARGIN;
   try {
     var blockBox = this.getCanvas().getBBox();
   } catch (e) {
     // Firefox has trouble with hidden elements (Bug 528969).
     return null;
   }
+  // Fix scale.
+  var contentWidth = blockBox.width * this.scale;
+  var contentHeight = blockBox.height * this.scale;
+  var contentX = blockBox.x * this.scale;
+  var contentY = blockBox.y * this.scale;
   if (this.scrollbar) {
     // Add a border around the content that is at least half a screenful wide.
     // Ensure border is wide enough that blocks can scroll over entire screen.
-    var MARGIN = 5;
-    var leftScroll = this.RTL ?
-        Blockly.Scrollbar.scrollbarThickness : 0;
-    var rightScroll = this.RTL ?
-        0 : Blockly.Scrollbar.scrollbarThickness;
-    var leftEdge = Math.min(blockBox.x - viewWidth / 2,
-        blockBox.x + blockBox.width - viewWidth - leftScroll + MARGIN);
-    var rightEdge = Math.max(blockBox.x + blockBox.width + viewWidth / 2,
-        blockBox.x + viewWidth + rightScroll - MARGIN);
-    var topEdge = Math.min(blockBox.y - viewHeight / 2,
-        blockBox.y + blockBox.height - viewHeight + MARGIN);
-    var bottomEdge = Math.max(blockBox.y + blockBox.height + viewHeight / 2,
-        blockBox.y + viewHeight + Blockly.Scrollbar.scrollbarThickness -
-        MARGIN);
+    var leftEdge = Math.min(contentX - viewWidth / 2,
+                            contentX + contentWidth - viewWidth);
+    var rightEdge = Math.max(contentX + contentWidth + viewWidth / 2,
+                             contentX + viewWidth);
+    var topEdge = Math.min(contentY - viewHeight / 2,
+                           contentY + contentHeight - viewHeight);
+    var bottomEdge = Math.max(contentY + contentHeight + viewHeight / 2,
+                              contentY + viewHeight);
   } else {
     var leftEdge = blockBox.x;
     var rightEdge = leftEdge + blockBox.width;
@@ -560,22 +591,10 @@ Blockly.setMainWorkspaceMetrics_ = function(xyRatio) {
   if (this.options.gridPattern) {
     this.options.gridPattern.setAttribute('x', x);
     this.options.gridPattern.setAttribute('y', y);
-  }
-};
-
-/**
- * Execute a command.  Generally, a command is the result of a user action
- * e.g., a click, drag or context menu selection.  Calling the cmdThunk function
- * through doCommand() allows us to capture information that can be used for
- * capabilities like undo (which is supported by the realtime collaboration
- * feature).
- * @param {function()} cmdThunk A function representing the command execution.
- */
-Blockly.doCommand = function(cmdThunk) {
-  if (Blockly.Realtime.isEnabled) {
-    Blockly.Realtime.doCommand(cmdThunk);
-  } else {
-    cmdThunk();
+    if (goog.userAgent.IE) {
+      // IE doesn't notice that the x/y offsets have changed.  Force an update.
+      this.updateGridPattern_();
+    }
   }
 };
 
@@ -608,4 +627,3 @@ if (!goog.global['Blockly']) {
 }
 goog.global['Blockly']['getMainWorkspace'] = Blockly.getMainWorkspace;
 goog.global['Blockly']['addChangeListener'] = Blockly.addChangeListener;
-goog.global['Blockly']['removeChangeListener'] = Blockly.removeChangeListener;
